requirements.txt
requests~=2.28.1
Flask~=2.1.2
python-dotenv~=0.20.0